# openapi.yaml
openapi: 3.1.0
info:
  title: Exotica ServiceTitan Platform — Accounting Reconciliation
  version: 1.0.0
  description: |
    Module 1 of the Exotica ServiceTitan Platform.

    Reconciles invoices and payments between a tenant's ServiceTitan account and
    their QuickBooks company, detects discrepancies, and produces AI-generated
    explanations and recommended actions.

    ## Architectural contract

    This spec is the ONLY interface between the platform core and any external
    consumer, including n8n workflows. External consumers never access the
    database directly. The product must function correctly with all external
    consumers removed — they are optional convenience layers, never dependencies.

    Consequences of that rule, which implementers must honour:

    - Scheduling, approval recording, and notification dispatch are implemented
      in the platform core, not by callers. Endpoints here expose and trigger
      them; they do not delegate them.
    - Every mutating endpoint is idempotent via `Idempotency-Key`.
    - Every read and write this platform performs against a tenant's ServiceTitan
      or QuickBooks account is recorded in the audit log and is retrievable
      through `/audit-events`.

    ## Versioning

    This document is versioned in Git and changes go through PR review.
    Additive changes (new optional fields, new endpoints) bump the minor version.
    Breaking changes require a new path prefix (`/api/v2/`), never an edit here.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /accounting/reconcile:
    post:
      tags: [Reconciliation]
      summary: Trigger a reconciliation run
      operationId: triggerReconciliation
      description: |
        Starts a reconciliation run for a tenant over a date range. Returns
        immediately with a `run_id`; the run itself is processed asynchronously.

        Only one run may be in progress per tenant at a time. A second request
        while a run is active returns 409 with the active run's id.

        The platform's own scheduler triggers runs on the tenant's configured
        cadence. This endpoint exists for on-demand runs and does not replace
        scheduled execution.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id:
                  type: string
                  format: uuid
                date_from:
                  type: string
                  format: date
                  description: >
                    Start of the period to reconcile. Defaults to 60 days before
                    `date_to`.
                date_to:
                  type: string
                  format: date
                  description: End of the period. Defaults to today.
                trigger_source:
                  type: string
                  enum: [manual, scheduled, api]
                  default: api
                  description: Recorded on the run for audit purposes.
            examples:
              onDemand:
                summary: On-demand run for the last 60 days
                value:
                  tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                  trigger_source: manual
              monthEnd:
                summary: Explicit month-end range
                value:
                  tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                  date_from: '2026-08-01'
                  date_to: '2026-08-31'
                  trigger_source: scheduled
      responses:
        '202':
          description: Run accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationRunAccepted'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: A run is already in progress for this tenant
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      active_run_id:
                        type: string
                        format: uuid
        '422':
          description: >
            Tenant is not ready to reconcile — for example ServiceTitan or
            QuickBooks credentials are missing or expired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /accounting/runs/{run_id}:
    get:
      tags: [Reconciliation]
      summary: Get a reconciliation run
      operationId: getReconciliationRun
      description: |
        Poll this endpoint after triggering a run. Terminal statuses are
        `succeeded`, `failed`, and `partial`. A `partial` run completed but could
        not fetch all data from one of the source systems; `source_errors`
        explains what was missed.
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: The run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconciliationRun' }
              examples:
                running:
                  summary: Still processing
                  value:
                    run_id: 7c1b9e22-5a4f-4c31-b0d8-2e9a7f60c114
                    tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                    status: running
                    trigger_source: scheduled
                    date_from: '2026-07-07'
                    date_to: '2026-09-05'
                    started_at: '2026-09-05T06:00:04Z'
                    completed_at: null
                    servicetitan_invoice_count: 45
                    quickbooks_invoice_count: null
                    matched_count: null
                    discrepancy_count: null
                    health_score: null
                    source_errors: []
                succeeded:
                  summary: Completed with discrepancies found
                  value:
                    run_id: 7c1b9e22-5a4f-4c31-b0d8-2e9a7f60c114
                    tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                    status: succeeded
                    trigger_source: scheduled
                    date_from: '2026-07-07'
                    date_to: '2026-09-05'
                    started_at: '2026-09-05T06:00:04Z'
                    completed_at: '2026-09-05T06:02:41Z'
                    servicetitan_invoice_count: 45
                    quickbooks_invoice_count: 42
                    matched_count: 42
                    discrepancy_count: 3
                    health_score: 0.9333
                    source_errors: []
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /accounting/runs/{run_id}/discrepancies:
    get:
      tags: [Reconciliation]
      summary: List discrepancies found by a run
      operationId: listRunDiscrepancies
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: severity
          in: query
          schema: { $ref: '#/components/schemas/Severity' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/DiscrepancyType' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of discrepancies
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Discrepancy' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────────────────────────────────────────────────────
  # Discrepancies
  # ─────────────────────────────────────────────────────────────────────────
  /accounting/discrepancies:
    get:
      tags: [Discrepancies]
      summary: List discrepancies across runs
      operationId: listDiscrepancies
      description: >
        Tenant-wide view. Use `status=open` to get everything currently
        outstanding regardless of which run detected it.
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DiscrepancyStatus' }
        - name: severity
          in: query
          schema: { $ref: '#/components/schemas/Severity' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/DiscrepancyType' }
        - name: min_amount
          in: query
          description: Filter to discrepancies whose absolute difference is at least this.
          schema: { type: string, description: Decimal string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of discrepancies
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Discrepancy' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /accounting/discrepancies/{discrepancy_id}:
    get:
      tags: [Discrepancies]
      summary: Get a single discrepancy
      operationId: getDiscrepancy
      parameters:
        - $ref: '#/components/parameters/DiscrepancyId'
      responses:
        '200':
          description: The discrepancy, including AI analysis if complete
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Discrepancy' }
              examples:
                underpayment:
                  summary: ServiceTitan shows paid, QuickBooks shows partial
                  value:
                    discrepancy_id: d41c8a90-77bb-4e0a-9c2f-1a5b3e88d902
                    run_id: 7c1b9e22-5a4f-4c31-b0d8-2e9a7f60c114
                    tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                    type: underpayment
                    severity: warning
                    status: open
                    servicetitan_invoice_id: 'INV-045'
                    quickbooks_invoice_id: '1042'
                    customer_name: Northside HVAC
                    servicetitan_amount: '5000.00'
                    quickbooks_amount: '3000.00'
                    difference: '2000.00'
                    currency: USD
                    detected_at: '2026-09-05T06:02:38Z'
                    analysis:
                      explanation: >-
                        ServiceTitan records two payments totalling $5,000 against
                        this invoice. QuickBooks reflects only the first $3,000
                        payment dated 2026-09-05. The second payment is most
                        likely sitting in Undeposited Funds or has not yet cleared.
                      likelihood: 0.85
                      recommended_action: >-
                        Check the QuickBooks Undeposited Funds account for a
                        $2,000 entry. If present, deposit it to the operating
                        account. If absent, confirm receipt with the customer.
                      requires_human_review: true
                      model: claude-sonnet-4-6
                      analysed_at: '2026-09-05T06:02:40Z'
                    approval: null
                    resolution: null
                missingInQuickbooks:
                  summary: Invoice exists in ServiceTitan only
                  value:
                    discrepancy_id: b7702e15-3c48-4f9d-8a61-0d4c7e2f1a66
                    run_id: 7c1b9e22-5a4f-4c31-b0d8-2e9a7f60c114
                    tenant_id: 3f9a1c40-2b6e-4d7a-9f11-8c2de5a41b03
                    type: missing_in_quickbooks
                    severity: error
                    status: awaiting_approval
                    servicetitan_invoice_id: 'INV-067'
                    quickbooks_invoice_id: null
                    customer_name: Bexley Plumbing
                    servicetitan_amount: '12400.00'
                    quickbooks_amount: null
                    difference: '12400.00'
                    currency: USD
                    detected_at: '2026-09-05T06:02:39Z'
                    analysis:
                      explanation: >-
                        This invoice was created in ServiceTitan on 2026-08-28 and
                        has no counterpart in QuickBooks over the reconciled
                        period. The customer record exists in both systems, so
                        this is most likely a sync gap rather than a data-entry
                        error.
                      likelihood: 0.72
                      recommended_action: >-
                        Create the corresponding sales invoice in QuickBooks for
                        $12,400 against Bexley Plumbing, dated 2026-08-28.
                      requires_human_review: true
                      model: claude-sonnet-4-6
                      analysed_at: '2026-09-05T06:02:41Z'
                    approval:
                      approval_id: 91a6f3c8-4e2d-4b70-a8f5-6c9b1d0e7f23
                      status: pending
                      requested_at: '2026-09-05T06:02:42Z'
                      reason: amount_above_threshold
                      threshold_amount: '10000.00'
                      decided_at: null
                      decided_by: null
                    resolution: null
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key for this logical operation. Replaying a
        request with the same key returns the original response and performs no
        additional work. Keys are retained for 24 hours.

        Required on every mutating endpoint. External callers retry; without
        this you will double-post approvals and double-trigger runs.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous response's `page.next_cursor`.
      schema: { type: string }
  responses:
    BadRequest:
      description: Request was malformed or failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token is valid but not authorised for this tenant or action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist or is not visible to this token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code, safe to branch on.
          examples: [validation_failed, run_in_progress, approval_already_decided]
        message:
          type: string
          description: Human-readable explanation. Wording may change; do not parse.
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
        request_id:
          type: string
          description: Correlates with server logs. Include when reporting problems.
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
...
    Severity:
      type: string
      description: |
        `info` — expected variance, no action needed.
        `warning` — probable timing or process gap, review recommended.
        `error` — data is inconsistent between systems, action required.
      enum: [info, warning, error]
    DiscrepancyType:
      type: string
      enum:
        - missing_in_quickbooks
        - missing_in_servicetitan
        - amount_mismatch
        - underpayment
        - overpayment
        - payment_count_mismatch
        - timing_mismatch
        - duplicate_invoice
...
    ReconciliationRunAccepted:
      type: object
      required: [run_id, status]
      properties:
        run_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RunStatus' }
        poll_url:
          type: string
          format: uri
          description: Convenience URL for polling this run.
        estimated_duration_seconds:
          type: integer
    ReconciliationRun:
      type: object
      required: [run_id, tenant_id, status, date_from, date_to, started_at]
      properties:
        run_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RunStatus' }
        trigger_source:
          type: string
          enum: [manual, scheduled, api]
        date_from: { type: string, format: date }
        date_to: { type: string, format: date }
        started_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        servicetitan_invoice_count: { type: [integer, 'null'] }
        quickbooks_invoice_count: { type: [integer, 'null'] }
        matched_count: { type: [integer, 'null'] }
        discrepancy_count: { type: [integer, 'null'] }
        health_score:
          type: [number, 'null']
          minimum: 0
          maximum: 1
          description: Matched invoices divided by total distinct invoices across both systems.
        source_errors:
          type: array
          description: Populated when status is `partial` or `failed`.
          items:
            type: object
            properties:
              system: { $ref: '#/components/schemas/SourceSystem' }
              error: { type: string }
              message: { type: string }
    Discrepancy:
      type: object
      required:
        [discrepancy_id, run_id, tenant_id, type, severity, status, difference, currency, detected_at]
      properties:
        discrepancy_id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/DiscrepancyType' }
        severity: { $ref: '#/components/schemas/Severity' }
        status: { $ref: '#/components/schemas/DiscrepancyStatus' }
        servicetitan_invoice_id: { type: [string, 'null'] }
        quickbooks_invoice_id: { type: [string, 'null'] }
        customer_name: { type: [string, 'null'] }
        servicetitan_amount:
          type: [string, 'null']
          description: Decimal string. Never a float — money is not a float.
        quickbooks_amount: { type: [string, 'null'] }
        difference:
          type: string
          description: Absolute difference as a decimal string.
        currency:
          type: string
          minLength: 3
          maxLength: 3
        detected_at: { type: string, format: date-time }
        analysis:
          oneOf:
            - $ref: '#/components/schemas/DiscrepancyAnalysis'
            - type: 'null'
          description: Null until AI analysis completes.
        approval:
          oneOf:
            - $ref: '#/components/schemas/Approval'
            - type: 'null'
        resolution:
          oneOf:
            - $ref: '#/components/schemas/Resolution'
            - type: 'null'
...